+++ inst/roxygen-examples/v_orient_temporal-example.R
@@ -24,6 +24,5 @@ skel <- pcalg::skeleton(
   alpha = 0.05,
   labels = vnames,
   method = "stable.fast",
-  fixedGaps = pc_cons$fixed_gaps,
-  fixedEdges = pc_cons$fixed_edges
+  fixedGaps = pc_cons$fixed_gaps
 )