+++ tests/testthat/test-constraint-based-methods.R
@@ -46,8 +46,8 @@ test_that("set_knowledge returns a new method and injects knowledge (all engines
       m2 <- set_knowledge(m, kn)
       expect_s3_class(m2, c(method_name, "disco_method", "function"))
       if (engine == "pcalg") {
-        # pcalg requires directed knowledge in both directions
-        expect_error(m2(num_data), "asymmetric edges")
+        # pcalg can't use directed knowledge
+        expect_warning(m2(num_data), "required edges")
       } else {
         expect_s3_class(m2(num_data), "Disco")
       }
@@ -77,10 +77,10 @@ test_that("disco() injects knowledge and validates method type (pc + fci)", {
       m <- do.call(reg$fn, c(list(engine = engine), args))
 
       if (engine == "pcalg") {
-        # pcalg requires directed knowledge in both directions
-        expect_error(
+        # pcalg can't use directed knowledge
+        expect_warning(
           disco(num_data, method = m, knowledge = kn),
-          "asymmetric edges"
+          "required edges"
         )
         next
       } else {